CHANGES TO APP.js + post is now STACK nav, one screen camera, other form. first part of form complete
