+++ screens/DishScreen.js
@@ -4,7 +4,6 @@ import { ScrollView } from 'react-native-gesture-handler';
 import { Text } from '../components/StyledText';
 
 export default function DishScreen() {
-  console.log("Dish Screen has mounted!");
   return (
     <ScrollView style={styles.container} contentContainerStyle={styles.contentContainer}>
